requirements.txt
fastapi
uvicorn[standard]
httpx
beautifulsoup4
tldextract
pillow
imagehash
piexif
